[skif] Context fully controls SkSpecialSurface creation
Previously FilterResult and the rest of image filtering was relying on
SkSpecialImage's makeSurface to create compatible SkSpecialSurfaces.
This doesn't work in the long term for two main reasons:

1. Image filters don't always require a non-null source image (i.e. they have no null children). Right now we can't detect that, draw to a layer
anyways, and then ignore the image except to use it to create special
surfaces. Null source images should be supportable, which requires the
skif::Context to know how to make surfaces as needed.
  - Related to this, transparency-affecting color filters can produce
non-null outputs from a null input. In this scenario they would also
rely on the Context's factory instead of requiring the source image to
be non-null.

2. We'd like to move away from SkSpecialImage entirely. This is a ways
out but when it's done, image creation will need som external means of
continuing to create compatible images and surfaces.
  - In the interim, FilterResult will actually support both
SkSpecialImage and SkImage as pixel data (since an SkImage can be
wrapped as an SkImageFilter).

Bug: skia:9279
Change-Id: I15acbf2ebb7b32bd54ef9f7e8002ec61efebe109
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/682876
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>