[Metal] Block out deprecated GPU family interface on newer SDKs.
MTLFeatureSet is deprecated with SDK 13.3 and later so we shouldn't
try to use it with those SDKs. MTLGPUFamily is available so this
shouldn't be an issue, and in the worst case we fall back to minimum
defaults.

Also updates our SDK versions, and renames Graphite defines to avoid
name collision.

Bug: chromium:1433578
Change-Id: I0c1c531884feac94224b7f6caf44e4e9c1f29925
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/684436
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>