Remove Ganesh and Graphite code from SkImage::subset
This makes both implementations more consistent and easier to
decouple if there is no GPU backend compiled in.

Graphite images are no longer uploaded as a texture-backed image
if a recorder is passed in and Lazy images are turned into
a raster image before being subset (instead of a texture-backed
image). In practice, this should only impact skp images
(which were further cleaned up in [1]).

In some cases, we want to be able to directly create a texture-
based subset, so a new static function does that
SkImages::SubsetTextureFrom. This currently has a trivial
implementation, but we might do something smarter, especially
with larger textures.

[1] https://skia-review.googlesource.com/c/skia/+/678656

Change-Id: I28f06a96fa451272c77a5fd06e97b2a4ad5b77b3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/671679
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>