Eliminate redundant constant-stores to the same range in SkRP.
Variable declarations without an initializer always set the value
to zero; this allows us to eliminate the zero-store if we
immediately write to the variable in the next statement.

(I tried similar checks in the other copy-to-slots functions, but
these attempts largely didn't pay off, so I left them as-is. No need
to pay for an optimization that didn't work.)

Change-Id: Ic79c14b873cfe2b38f3f6b1b246172f1ce8ad1d6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/684676
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>