Remove unused AVX512 opts code
These were all contributed by Intel:
  https://skia-review.googlesource.com/c/skia/+/284528
  https://skia-review.googlesource.com/c/skia/+/290436

Later, we decided to remove them from the SKX opts translation unit:
  https://skia-review.googlesource.com/c/skia/+/313809
  https://skia-review.googlesource.com/c/skia/+/313865

This was because of concerns over clock speed reduction. At this point,
I suspect the clock-speed reduction is not that big of an issue, but we
haven't had this code active for several years, and it's really unlikely
that anyone cares.

It's technically possible that someone is compiling all of
Skia with -march=skylake-avx512, but I doubt it.

Change-Id: I3e72d36a43ff06db7b52980f678c165aa38ba5ac
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/684579
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>