Remove AVX512 (SKX) logic from RasterPipeline_opts
This didn't actually do anything (we only build this file with AVX2, and
all of the checks were covering both architectures). Mostly just trying
to trim back the number of places that mention SKX, which is likely to
disappear completely in a little bit.

Change-Id: Ifea572b37b293ca59dab271dd6e9f803c588352d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/684816
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>