Remove SkSpecialImage::makeTightSurface
The sole use had been in SkTileImageFilter which was using it to create
a non-special SkImage that could use kRepeat with HW tiling.

Now that SkSpecialImage exposes `asShader(SkTileMode)` that operates on
the subset of valid data within the special image, this extra step is
no longer necessary.

It still has to draw to another image if the tiling boundary wasn't
completely fulfilled by the input filter but that can now use
SkSpecialSurface.

Bug: skia:9279
Change-Id: Ifd9bad3b3906f658708ce38badde4bf8f077e889
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/683376
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>