[graphite] Track buffer access mask and use for GPU->CPU transfer.
Adds a new field to VulkanBuffer for tracking the current accessMask
bits, and code to execute memory barriers when necessary for certain
accessMask transitions. Currently only sets the access mask when
transferring GPU texture data to a mappable buffer and reading it
back on the CPU. This is the only case that needs a barrier at the
moment.

Change-Id: If2143b93023e7ec0776f925328006ef802a480ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/686316
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>