Remove abs_float SkRP ops.
In this CL, we replace each abs(float) operation with a bitwise-and
against 0x7FFFFFFF. In the case of multi-slot abs operations, this
costs an extra instruction. In the followup CL, we add a multi-slot
immediate-mode bitwise-and op, which reclaims these losses and is
a more flexible op.

Change-Id: I85ca34ccfcf045dabdbb41d38e5766c850fae489
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/687641
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>