[skif] Remove use of deprecated Magnifier factory within Skia
I deleted one old GM that didn't seem to be adding much value, and
adjusted the others to sort-of reconstruct the original lens bounds
and zoom amount. It's not perfect so I expect the gold images to
qualitatively change.

Updated the image filter fuzzer to use the new magnifier impl.

Bug: chromium:1425773
Bug: b/280470437
Change-Id: Id83fc9044056d1cb6ec78d81ced6c93cdf35b7be
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/689636
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>