Add multi-slot immediate bitwise-and.
This reclaims the losses from the previous CL, and replaces a
single-use instruction (abs) with a more flexible one
(bitwise-and). Since we use bitwise-and to cast from boolean to
float and int, this improves codegen for vector-boolean casts.

Change-Id: Ifedd91c876190b2151cb8de91f2fa3b35f775b0a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/687643
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>