[skif] Re-implement Magnifier image filter
The new version is written as an entirely new subclass of
SkImageFilter_Base since chromium currently relies on the legacy
magnifier implementation.

Logic that was in chrome to adapt to cropping and calculate the source
content has been moved into onFilterImage() and wrapped in coordinate
space-tagging types. The new implementation also participates in
bounds tracking.

Lastly, this consolidates the implementation down to use a runtime
shader for both CPU and GPU. This means it *should* work in Graphite.
As more filters are updated to this pattern, some of the boilerplate
in this CL will be migrated into FilterResult. The runtime shader is
effectively the same as the one used for the legacy ganesh impl, but
it has had its uniforms and variable names renamed to be clearer.

Bug: chromium:1425773
Bug: b/280470437
Change-Id: I9c31203eeb7bbcd717e45bc466fd79a56cd1076d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/689396
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>