Generate simpler code for sk_Clockwise in SPIR-V.
While working on a no-DSL refactor of this code, I substituted a
logical-xor for a ternary expression and was surprised at the
substantial ripple effect in the generated code. Seems fine, but I
decided to land this as a standalone CL just in case.

Change-Id: I54445798caccc7098603d52e907a40721c83628c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/694816
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>