Add SkNoDestructor<T> class to Skia.
This class is adapted from Chromium, and is BSD-licensed:
https://crsrc.org/c/base/no_destructor.h

`static SkNoDestructor<T> foo(...);` inside a function approximates
a global singleton, but is safe from ordering concerns at startup or
shutdown. It is also faster than allocating an object on the heap.
(We used to consider allocation from the heap as a sufficiently-
good solution, but apparently this leads to measurable overhead
when everyone does it:
http://go/totw/110#the-fix-safe-initialization-no-destruction)

I've added it to SkSL in places where we currently allocate.

Change-Id: I2f1c6cf915c6f1294242951206a9d40e640fa4cc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/694624
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>