[sksl] Rename Program::Inputs to Program::Interface
The motivation for this is to store output settings in the struct,
namely for a secondary color output, without the name being misleading.

The name "interface" is avoided in headers and D3D source, since it is
defined as a macro in some Windows system headers that are included
along with D3D headers.

Change-Id: I8f91fde9676ce042b61631b12d226f3bde71fc01
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/695278
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>