Remove DSLField type.
This type simply mirrored the native SkSL field type, so it's
wasteful to build up an array of DSLFields, just to copy it into
a parallel array of SkSL::Fields. I've also updated the APIs to
use TArray consistently throughout; before, it was a mix of vectors,
SkSpans, and TArrays, which made it difficult to use move()
consistently.

Change-Id: I7cb895d6622467c9943da55d5f5bed339f0f914f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/696597
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>