Make IWYU build skip linking
IWYU is a compile-time check, so we can make it go a little bit
faster by skipping the linking of an archive or executable.

Since we are skipping the linking, we can skip the actual compile
of files we are enforcing IWYU on. This means the output of an
IWYU build cannot be used for anything, but since we had not been
doing that (e.g. not running tests on it), nothing is lost and
about 30 seconds on a local build of the //:skia_public target
is saved.

Change-Id: I68472f62b60f0c2a079b2ac80c0cfa2f5a7d1d74
Bug: skia:13052
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/696539
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>