Remove SkVM from the SkSL test matrix.
Previously, we had `RP` which indicated SkRP support, and `VM`
indicating SkVM support. Now we have `CPU`, which indicates
a SkRuntimeEffect-compatible (ES2 only) shader, and `ES3` which
indicates a shader that is not ES2-compatible, but one that SkRP can
pass if we call `MakeRasterPipelineProgram` directly.

I experimented with allowing SkRuntimeEffect to honor the
`maxVersionAllowed` option in software, but this change ended
up having a lot of ripple effects into SkCapabilities, so I
shelved it for now.

Change-Id: I5c6abd4217be6463f104eddeb882be6e4fcd4453
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/695307
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>