Remove swap and assign from DSLExpression.
`assign` was no longer used anywhere.

`swap` was used in the Parser because DSLExpression used to
have a custom operator= overload; this meant that it could not
be assigned-to in the parser directly. However, this overload is
long gone, and these swaps are no longer needed. Instead, we can
just use regular move-assignment.

The basic set of constructors were cleaned up to just use
`= default` and `= delete`.

Change-Id: I61a8fc41185278e04c3bf3db6cd3a4995056dd15
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/698238
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Nolan Scobie <nscobie@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>