Create Protected EGL contexts when possible on Android
Plumbing a command line parameter down to the EGL level is a bit inconvenient (esp. for units tests where we often create GrContextFactories on the stack). Having a global variable actually seems to be the cleanest method :(. Note that we need to create a Protected EGL context in order for GL to access its Protected capabilities.


Bug: skia:14298
Change-Id: Ie0ea71810371801382c84b9664807e522a058661
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/695617
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>