[skif] Shore up getOutputLayerBounds() for old filter definitions
The new bounds API and implementations assume that if a node affects
transparent black, it will return an infinite or very large bounds for
its output bounds. This can then be intersected or detected using
bounds math.

Old implementations relied on this happening in getOutputBounds(),
but that's only valid as long as the old filter was the root node.
With SkColorFilterImageFilter being ported, that is often not the
case with SVGs produced by chromium.

Bug: skia:9282
Change-Id: Iefa3f684c3a1878100c38561e86f61ff59a6bf58
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/699097
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>