[vello] Support compiling the vello dependencies using Bazel
Added support for compiling the vello_cpp crate and its external
dependencies using bazel's built-in support for Rust/Cargo. The primary
consumer of the vello code is Graphite which currently only compiles
with GN. This is supported using a custom GN action that invokes bazel
to build the Rust dependencies of the //third_party:vello GN target.

Bug: b/241583094

Change-Id: I7ac04299fc76f91b81dc88e3e7f645989b95e0d2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/699179
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>