Reorganize DirectWrite glyph representation selection
A number of glyph representations were checked for existence first and
then there was an attempt to retrieve them if the test came back
positive. Instead, just try to get the representation directly and
inline any necessary checks. This is both faster and easier to
understand. The "check first" strategy existed so that both the metric
and image code would use the same tests for which representation to use.
This is no longer required as the more robust ScalerContextBits are now
being used to communicate the representation to use directly.

Change-Id: I2a3948dace2d8bd2711551c8c657ea386f9d2393
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/699777
Reviewed-by: Herb Derby <herb@google.com>