[graphite] Fix YV12 colors.
The previous code assumed that the plane order matched the color
channel order, but YV12 stores the data in Y_V_U order. The
YUVALocations array provides the mapping for a given YUVA channel
to a plane index and component index in that plane to get that
channel value, so we should iterate through that to set up the
texture and channelSelect values.

Bug: b/274150838
Change-Id: I42e69c01bab3c0da0cd3e8b0a0d2843b7fbc684b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/699598
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>