Regularize DirectWrite format handling
Originally the DirectWrite generateMetrics and generateImage only needed
to handle glyphs represented by the masks that DirectWrite would produce
from outlines in the font. So the code for this case lived directly
in generateMetrics and generateImage. Now COLR, SVG, PNG, and SkPath
fallback are all supported in addition to the original DirectWrite
produced masks, with COLRv1 on the way. Re-organize this code so that
the top level generateMetrics and generateImage are solely responsible
for picking which representation to use and putting the implementation
of each representation into a regularly named generate*Metrics and
generate*Image method.

Change-Id: Iac022f1e9208f726fc57a71c8a9ccfec9a85f38d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/699778
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>