[skif] Don't evaluate empty-source filters for SkNoDrawCanvases
With OOP-R, filters that didn't rely on a source image would now be
evaluated with a raster context only to have the SkNoPixelsDevice's
drawSpecla() discard the output of the filter.

SkCanvas::internalRestore() has an equivalent check, but it was not
carried to this up-front filtering path.

Bug: chromium:1446547
Change-Id: I33d20d484c25d77ddbc4b463ab909e6f426dec01
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/700396
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>