[bazel] BazelTest-* task schema: Add support for precompiled Android tests.
This CL makes changes to the bazel_build.go task driver to optionally store a subset of //bazel-bin into CAS, and adds a new task driver (and example task) that reads said files as CAS inputs and runs an arbitrary command involving those files.

The goal is to compile an Android test on a GCE machine, and run the precompiled test on a Raspberry Pi attached to an Android device. The reason for this split is that we don't want to run Bazel on a Raspberry Pi due to its constrained resources.

Recommended review order:
- jobs.json
- gen_tasks_logic.go
- bazel_build.go
- bazel_test_precompiled.go
- tasks.json
- Everything else

Bug: skia:14227
Change-Id: I7d4b7ab6bf43f5c07cbd394c4cf944c2449a6d62
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/694625
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>