[bazel] //tests/adb_test.bzl: Rewrite test runner in Go.
This CL rewrites the adb_test test runner in Go. In order to run the Go test runner on a Raspberry Pi, we must cross-compile it for linux_arm64. This is done via a transition rule that is controlled via a new --adb_test_runner_target_platform flag. When the flag is unset, the transition defaults to the host platform, meaning that Skia developers won't need to set this flag when running tests locally against an Android device attached to their workstation (or by port-forwarding a remote ADB server).

On the CI side, this CL changes the relevant BazelBuild-* tasks so that tests are built with --adb_test_runner_target_platform=linux_arm64.

Bug: skia:14227
Change-Id: I9a8bfcc147c27282be6e836e73ff598a3dcb3168
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/699793
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>