[skif] Remove dedicated SkDropShadowImageFilter implementation
Offsets and color filters now use FilterResult to compose together
without producing intermediate renderpasses. This means that the
image filter sub graph that describes a drop shadow should have an
equivalent number of renderpasses as this dedicated image (which
internally applied an offset during drawing, a color filter and
recursed with a blur image filter).

Besides functions for creating the graph of image filters, the
majority of this CL is adjusting the SKP deserialization so that
SKPs that referenced the old SkDropShadowImageFilter CreateProc can
be reconstructed until the min picture version has moved past.

Bug: skia:9283
Bug: b/263133333
Change-Id: I75ae0772c254d0633b8ca4621d63a48ed683470f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/688136
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>