Update all versions of Bazelisk and add one for Intel Macs
The impetus for this is the follow-on CL which uses Bazel in
the Skia CI on Intel-based Macs (for the first time [?]).

We previously had just a "bazelisk" package, which was implicitly
for x64 Linux. All other versions had a suffix for the OS and
architecture. This makes Linux look like all the other ones,
adds a helper function to pick the right one, updates all
asset scripts to verify a sha256 hash, *and* updates all of them
to be the latest bazelisk version (1.17.0) for consistency.

I don't know if the mac_arm64 version is actually used, but
the win_amd64 one is by skcms.

I plan to delete the bazel and bazelisk asset folders (but keep
all the CIPD stuff in case for backwards compatability). For
now, I just removed the create.py scripts and added a README.md.
I think this requires a change in the Skia Infra repo.

Change-Id: I7bbd73e36cba6909d7d0a15168b83a188b6f6621
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/701436
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>