[graphite] Clean out ProxyCache before unrefing Resources' cacheref.
We had an ordering issue when shutting down the ResourceCache. The
ProxyCache has a usageRef on some textures but was being torn down
after the Resources in the cache had their cacheRefs decremented,
which led to the textures not being deleted. The solution is to
clear out the ProxyCache to clear the usageRef, so that decrementing
the cacheRef will delete the texture.

Bug: b/277583609
Change-Id: Ie9546950536c2517d8139a9f17ea10eee4f7e196
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/702137
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>