[graphite] Add VulkanImageView support to VulkanTexture.
VulkanImageViews are constructed from the VulkanTexture using their
usage, either as color input to a shader or as an attachment on a
renderpass. They are then added to an array that can be searched
through for future uses. For now this is all we need, but this gives
us the flexibility to add additional parameters and a more complicated
caching system in the future if needed.

Bug: b/283822171
Change-Id: Iff86fc2e65bb0775e5bc1438604b6f96c4f82feb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/701521
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>