[Fontations] Don't assert scaler understands all outlines
The scaler will not necessarily be able to understand the font file. As
generatePath() in SkScalerContext can fail, we can pass that error
up. The previous assert!() macro was too strict and at the moment causes
a panic on CFF outlines or malformed fonts.

Bug: skia:14259
Change-Id: Ic346033d2c68f7cea4cdc9d3d446d09f74a15dbc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/703516
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Dominik Röttsches <drott@chromium.org>