[Fontations] Remove middle gen/ path extension for cxxbridge output
Inserting gen/ is not compatible with how Blink buildfiles generate
cxxbridge output and leads to requiring #include statements with
different paths in the Skia and in the Blink context.

Blink's GN rules for cxxbridge output the generated
files (as other Blink generated files) to a top-level gen/ directory in
the GN output directory, then add these paths as include path.

The path pointing to the source file in the #include directive in the
C++ file should a) be the same in Skia and Blink contexts and b) should
look identical to the path to the Rust source from which cxxbridge
generated this file. To that effect, remove the "gen/" path extension in
the run_cxxbridge_cmd rule.

Fixed: skia:14332
Change-Id: Ia3246f6a6a7073aa9f5d731003cf354b345e2459
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/703856
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>