SkJpegMetadataDecoder: Add Exif, ICC, and MPF support
Add function to retrieve Exif and ICC metadata, and a method to query
if a gainmap may be present (to early-out the more expensive call that
must be done to extract the gainmap).

Update read_metadata to use the helper function marker_has_signature,
and to also support padding after the signature (which Exif requires).

Change SkJpegCodec to use SkJpegMetadataDecoder for Exif and ICC
metadata extraction.

Bug: 1444342
Change-Id: I98da5036e8c1e27a54c8522d1de4abe4e82338de
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/694939
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>