Avoid taking the address of a vector-component in WGSL.
We can't take the address of a swizzle in WGSL. Instead, we take the
address of the inner expression and tack on the swizzle to the
returned expression. i.e., getting a scratch pointer to `foo.xy` would
take the address of `foo` and return `(*_skTemp123).xy` as the
substitute expression.

Bug: skia:14082
Change-Id: I1e1a8ad12ce333fe219d3dd3f116252ad5fe1d38
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/705233
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>