[infra] temporarily remove RadeonHD8870M-x86_64-Debug-All-Graphite_Metal from CQ
There was only one machine capable of running this (build336-a9) which
is in the process of moving to SFO36 as part of http://b/284341952.

Remove from CQ to avoid CLs getting hung up waiting for this
builder. Will return to the CQ when the host is back in
the swarm as mac-204-e505.

Change-Id: Ie9daf964bbacefb0e64e2ab5ffe97491542f1630
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/705982
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>