Update gn_to_cmake.py for Python 3
WriteActionForEach does a divide. In Python2, which was what was
being used when this was written, `int / int -> int`. In Python3 `int /
int -> float`. It could be possible to use `//` in Python3, but
converting to `int` does the same thing for the positive integers used
here.

Bug: skia:14349
Change-Id: I5f1693edb18db9768041c60021b43a8e239a1f1a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/706379
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>