[skif] Apply decal at layer resolution
This uses a runtime effect for now, but I think that there's a good
chance for utility SkShaders that handle tile modes on a float interval
to be added for some later image filter work, in which case hopefully
this can be updated to share the shader and have a dedicated raster
pipeline perhaps.

The gist is that if a deferred transform on a FilterResult would cause
decal sampling to visually be larger than the 1px ramp you'd expect
from AA at the layer resolution, then we do something extra to fix it.
This keeps results more consistent between wrapping as a shader or
drawing the image directly (which is "decal", but actually clamp
sampled and geometrically clipped to the image bounds).

If the image's transform does not preserve rectangles, we can't
just map the decal bounds to in front of the transform so the
logic splits the transform into a pre-decal and post-decal step.

Bug: skia:9282
Change-Id: Ia40d384b851a0b1717e82953ba78fad615a5a9db
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/706439
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>