[graphite] Support atlas draws
Device::drawGeometry() now supports scheduling an atlas draw when
rendering paths. Device::chooseRenderer() has been enhanced to return
both a Renderer (for the render pass) and a PathAtlas (which abstracts
the atlas rendering method).

Currently chooseRenderer() always returns the AtlasShape renderer and
the ComputePathAtlas if:
   1. The shape is complex and doesn't match a specialized renderer
      (such as AnalyticRRect, vertices, subrun data); and
   2. Vello shaders were enabled in the build; and
   3. The render does not specifically require hardware MSAA; and
   4. Compute shaders are supported.

This currently implicitly leaves out clip draws which currently require
hardware MSAA. We will need to come up with a reasonable approach for
when to route shapes to the atlas for clipping as atlas clipping may be
unnecessary it has to be combined with depth-buffer clips.

The atlas renderer selection criteria will also likely need to be fine
tuned, for example to make more effective use of available atlas space
for shapes that really benefit from it.

Bug: b/280927575
Change-Id: I1ba3842e00db0df86553a559dc4217182c0a8a9f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/707056
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>