Do not return infinities as roots
The Cubic solver uses the quadratic solver when the x^3
coefficient is very close to 0. The quadratic solver was
returning infinity for the roots because the roots are
not in the range of doubles. The cubic code can't deal
with infinity so change to return 0 roots.

Bug: oss-fuzz:59626
Change-Id: I29d0aa21a1bb7daf53237a17b62e67e8c13f29ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/708117
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>