Rename OutParamsTricky to OutParamsDoubleSwizzle.
The original "tricky" part of this test was an inout param that was
swizzled at the call site, then swizzled again during the assignment.
This intent was muted somewhat because the inner lvalue was swizzed
as `.xy`, which we later learned to optimize away entirely since
it's an identity swizzle.

This CL updates the swizzle-assignment line slightly; it has the same
effect, but the swizzle is now on the lvalue side. It also renames
the test from "OutParamsTricky" to "OutParamsDoubleSwizzle" in order
to better convey the trickiness at stake.

Change-Id: I94b9ee358975b8092b04aa67b8ca278e4c327ec0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/708118
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>